+++ src/tests/test_typeconversions.py
@@ -239,38 +239,8 @@ def test_error_from_js(selenium):
 
 
 def test_error_from_python(selenium):
-    selenium.run("err = Exception('hello there?');")
+    selenium.run("err = Exception('hello there?')")
     assert_py_to_js_to_py(selenium, "err")
-
-
-def test_jsproxy_attribute_error(selenium):
-    selenium.run_js(
-        """
-        class Point {
-            constructor(x, y) {
-                this.x = x;
-                this.y = y;
-            }
-        }
-        window.point = new Point(42, 43)
-        """
-    )
-    selenium.run(
-        """
-        from js import point
-        assert point.y == 43
-        """
-    )
-
-    msg = "AttributeError: z"
-    with pytest.raises(WebDriverException, match=msg):
-        selenium.run("point.z")
-
-    selenium.run("del point.y")
-    msg = "AttributeError: y"
-    with pytest.raises(WebDriverException, match=msg):
-        selenium.run("point.y")
-    assert selenium.run_js("return point.y;") is None
 
 
 def test_javascript_error(selenium):
