+++ src/type_conversion/hiwire.c
@@ -317,6 +317,11 @@ EM_JS(int, hiwire_is_function, (int idobj), {
   // clang-format on
 });
 
+EM_JS(int, hiwire_is_pyproxy, (int idobj), {
+  return Module.hiwire.new_value(
+    Module.PyProxy.isPyProxy(Module.hiwire.get_value(idobj)));
+});
+
 EM_JS(int, hiwire_bind, (int idobj, int idthis), {
   // clang-format off
   return Module.hiwire.new_value(Module.hiwire.get_value(idobj).bind(Module.hiwire.get_value(idthis)));
